refactor: extract lunar markers to InkScrollView extension to fix lint
Move LunarMarkerDot, LunarMarker, computeLunarMarkers,
findLunarEvents, refinePeak, and interpolatePosition to
InkScrollView+LunarMarkers.swift to bring InkScrollView struct
body under the 700-line SwiftLint limit.
